refactor: move fetching actions into actions_gateway
